API: renaming modules and their test modules
[CHANGED] the modules are not just about db's, so renaming
     them to reflect their broader scope.
